feat: remove sidebar options, go for history only and pin
