+++ apps/array/src/renderer/features/sidebar/components/HistoryView.tsx
@@ -1,12 +1,18 @@
 import { useTaskExecutionStore } from "@features/task-detail/stores/taskExecutionStore";
 import { Button, Flex } from "@radix-ui/themes";
 import { useWorkspaceStore } from "@/renderer/features/workspace/stores/workspaceStore";
-import type { HistoryData, HistoryTaskData } from "../hooks/useSidebarData";
+import type {
+  HistoryData,
+  HistoryTaskData,
+  PinnedData,
+  TaskData,
+} from "../hooks/useSidebarData";
 import { useSidebarStore } from "../stores/sidebarStore";
 import { TaskItem } from "./items/TaskItem";
 
 interface HistoryViewProps {
   historyData: HistoryData;
+  pinnedData: PinnedData;
   activeTaskId: string | null;
   onTaskClick: (taskId: string) => void;
   onTaskContextMenu: (taskId: string, e: React.MouseEvent) => void;
@@ -65,8 +71,52 @@ function HistoryTaskItem({
   );
 }
 
+interface PinnedTaskItemProps {
+  task: TaskData;
+  isActive: boolean;
+  onClick: () => void;
+  onContextMenu: (e: React.MouseEvent) => void;
+  onDelete: () => void;
+  onTogglePin: () => void;
+}
+
+function PinnedTaskItem({
+  task,
+  isActive,
+  onClick,
+  onContextMenu,
+  onDelete,
+  onTogglePin,
+}: PinnedTaskItemProps) {
+  const workspaces = useWorkspaceStore.use.workspaces();
+  const taskStates = useTaskExecutionStore((state) => state.taskStates);
+
+  const workspace = workspaces[task.id];
+  const taskState = taskStates[task.id];
+
+  return (
+    <TaskItem
+      id={task.id}
+      label={task.title}
+      isActive={isActive}
+      worktreeName={workspace?.worktreeName ?? undefined}
+      worktreePath={workspace?.worktreePath ?? workspace?.folderPath}
+      workspaceMode={taskState?.workspaceMode}
+      lastActivityAt={task.lastActivityAt}
+      isGenerating={task.isGenerating}
+      isUnread={task.isUnread}
+      isPinned={task.isPinned}
+      onClick={onClick}
+      onContextMenu={onContextMenu}
+      onDelete={onDelete}
+      onTogglePin={onTogglePin}
+    />
+  );
+}
+
 export function HistoryView({
   historyData,
+  pinnedData,
   activeTaskId,
   onTaskClick,
   onTaskContextMenu,
@@ -76,11 +126,32 @@ export function HistoryView({
   const loadMoreHistory = useSidebarStore((state) => state.loadMoreHistory);
   const { activeTasks, recentTasks, hasMore } = historyData;
 
+  const hasPinnedTasks = pinnedData.tasks.length > 0;
   const hasActiveTasks = activeTasks.length > 0;
   const hasRecentTasks = recentTasks.length > 0;
 
   return (
     <Flex direction="column">
+      {hasPinnedTasks && (
+        <>
+          <HistorySectionLabel label="Pinned" />
+          {pinnedData.tasks.map((task) => (
+            <PinnedTaskItem
+              key={task.id}
+              task={task}
+              isActive={activeTaskId === task.id}
+              onClick={() => onTaskClick(task.id)}
+              onContextMenu={(e) => onTaskContextMenu(task.id, e)}
+              onDelete={() => onTaskDelete(task.id)}
+              onTogglePin={() => onTaskTogglePin(task.id)}
+            />
+          ))}
+          {(hasActiveTasks || hasRecentTasks) && (
+            <div className="mx-2 my-2 border-gray-6 border-t" />
+          )}
+        </>
+      )}
+
       {hasActiveTasks && (
         <>
           <HistorySectionLabel label="Active" />
