+++ apps/array/src/renderer/features/sidebar/stores/sidebarStore.ts
@@ -1,8 +1,6 @@
 import { create } from "zustand";
 import { persist } from "zustand/middleware";
 
-export type SidebarViewMode = "folders" | "history" | "pinned";
-
 interface SidebarStoreState {
   open: boolean;
   hasUserSetOpen: boolean;
@@ -10,7 +8,6 @@ interface SidebarStoreState {
   isResizing: boolean;
   collapsedSections: Set<string>;
   folderOrder: string[];
-  viewMode: SidebarViewMode;
   historyVisibleCount: number;
 }
 
@@ -24,7 +21,6 @@ interface SidebarStoreActions {
   reorderFolders: (fromIndex: number, toIndex: number) => void;
   setFolderOrder: (order: string[]) => void;
   syncFolderOrder: (folderIds: string[]) => void;
-  setViewMode: (mode: SidebarViewMode) => void;
   loadMoreHistory: () => void;
   resetHistoryVisibleCount: () => void;
 }
@@ -40,7 +36,6 @@ export const useSidebarStore = create<SidebarStore>()(
       isResizing: false,
       collapsedSections: new Set<string>(),
       folderOrder: [],
-      viewMode: "history" as SidebarViewMode,
       historyVisibleCount: 25,
       setOpen: (open) => set({ open, hasUserSetOpen: true }),
       setOpenAuto: (open) =>
@@ -83,7 +78,6 @@ export const useSidebarStore = create<SidebarStore>()(
           }
           return state;
         }),
-      setViewMode: (mode) => set({ viewMode: mode }),
       loadMoreHistory: () =>
         set((state) => ({
           historyVisibleCount: state.historyVisibleCount + 25,
@@ -98,7 +92,6 @@ export const useSidebarStore = create<SidebarStore>()(
         width: state.width,
         collapsedSections: Array.from(state.collapsedSections),
         folderOrder: state.folderOrder,
-        viewMode: state.viewMode,
         historyVisibleCount: state.historyVisibleCount,
       }),
       merge: (persisted, current) => {
@@ -108,7 +101,6 @@ export const useSidebarStore = create<SidebarStore>()(
           width?: number;
           collapsedSections?: string[];
           folderOrder?: string[];
-          viewMode?: SidebarViewMode;
           historyVisibleCount?: number;
         };
         return {
@@ -119,7 +111,6 @@ export const useSidebarStore = create<SidebarStore>()(
           width: persistedState.width ?? current.width,
           collapsedSections: new Set(persistedState.collapsedSections ?? []),
           folderOrder: persistedState.folderOrder ?? [],
-          viewMode: persistedState.viewMode ?? current.viewMode,
           historyVisibleCount:
             persistedState.historyVisibleCount ?? current.historyVisibleCount,
         };
